MCS2023_baseline/requirements.txt
numpy==1.22.4
pandas==1.4.4
opencv-python==4.5.5.64
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
PyYAML==5.4.1
scikit-learn==1.1.1
scipy==1.8.1